\name{exampleArrayData}
\docType{data}
\alias{exampleArrayData}
\title{SNP Array Data to Use for Gene Set Analysis}
\description{
	This dataset provides simulated association p-values for an example set of SNP rsIDs, including their chromosome and position along the chromosome. The data in this file are all SNPs that lie on chromosome 1.
}
\usage{exampleArrayData}
\format{A data.frame containing 119,487 rows and the following 4 columns:
\tabular{ll}{
	P    \tab simulated p-values\cr
	SNP  \tab SNP rsIDs\cr
	Position \tab SNP position\cr
	chromosome \tab SNP chromosome
}
}
\keyword{datasets}