\name{geneSets}
\docType{data}
\alias{geneSets}
\title{Gene Set Collection object}
\description{
	This gene set collection object provides four gene sets of varying sizes to be used as examples of how to run the \code{cpvSNP} package functions.
}
\usage{geneSets}
\format{A \code{GeneSetCollection} object of length 4. The values stored in the \code{geneIds} slot are Entrez ids. Gene sets 1-3 were randomly selected with sizes 100, 200 and 50, respectively. The fourth GeneSet holds genes in which the simulated p-values for corresponding SNPs are significant.}
\keyword{datasets}