% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotPvals}
\alias{plotPvals}
\title{Create a Plot of P-Values Against Number of SNPs
per Set for All Sets in a Collection}
\usage{
plotPvals(set, title = "", xlab = "SNPs per Gene Set", ylab = "PValue",
  ...)
}
\arguments{
\item{set}{An object of class \code{GeneSetResultCollection}.}

\item{title}{A character string containing the title of the plot.
Default is "", a blank title.}

\item{xlab}{A character vector with the label for the x-axis on
the plot. Default is \code{SNPs per Gene Set}.}

\item{ylab}{A character vector holding the label for the y-axis
on the plot. Default is \code{PValue}.}

\item{...}{Further arguments to be passed to the plotting methods,
such as graphical parameters.}
}
\value{
Creates a plot.
}
\description{
Plots association test p-values against number of SNPs per
gene set.
}
\details{
Creates a plot of all p-values against the number of
SNPs in the gene set, for all sets within a collection.
}
\author{
Caitlin McHugh \email{mchughc@uw.edu}
}

