% Generated by roxygen2 (4.0.1): do not edit by hand
\name{simulate_chisq}
\alias{simulate_chisq}
\title{Simulate a Chi-squared Statistic from a Distribution
with a Specified Correlation Structure}
\usage{
simulate_chisq(corr_matrix, num_sims)
}
\arguments{
\item{corr_matrix}{A square matrix of correlations.}

\item{num_sims}{An integer value for the number of simulations
to be performed.}
}
\value{
A numeric value corresponding to a chi-squared
statistic simulated from a chi-squared-like distribution with
specified correlation structure and degrees of freedom
corresponding to the number of rows in the correlation matrix.
}
\description{
Simulates a value from a chi-squared distribution with a
specified correlation structure.
}
\details{
This function simulates a value from a chi-squared
distribution with a specified correlation structure and degrees
of freedom corresponding to the number of rows in the
correlation matrix.
}
\author{
Caitlin McHugh \email{mchughc@uw.edu}
}

